refactor: export all public types
