+++ src/audio/index.ts
@@ -1,0 +1,13 @@
+export {
+	AudioPlayer,
+	AudioPlayerStatus,
+	NoSubscriberBehaviour,
+	createAudioPlayer
+} from './AudioPlayer';
+
+export {
+	AudioResource,
+	createAudioResource
+} from './AudioResource';
+
+export { PlayerSubscription } from './PlayerSubscription';
