+++ src/util/index.ts
@@ -1,0 +1,3 @@
+import generateDependencyReport from "./generateDependencyReport";
+
+export * as generateDependencyReport from './generateDependencyReport';
